Audit the committed gold bytes: dev is 79% contaminated, test is clean
Adds scripts/audit_gold.py (make audit) plus tests that read the COMMITTED
bytes rather than regenerating them. That distinction is the point: the ADR
0011 clock defect survived a green suite for weeks because every test rebuilt
the data the same wrong way. These assert properties that must hold however
the data was produced.

Two defects found.

DEV IS UNUSABLE FOR MODEL SELECTION. 150 of 189 dev records (79.4%) appear
verbatim in data/train.jsonl. The data engine was seeded from dev, and the
dedup pass was handed the test split to check leakage against, so dev
contamination was never detected -- train.meta.json's "removed_leakage: 11"
counted only what it was asked to look for. Selecting a config on dev would be
scoring memorised training text and would reliably pick the most overfit
option. This blocks WP-3, whose plan was explicitly "sweep on dev only";
roadmap updated and WP-0d added to build a clean split under a new seed.
Pinned as an xfail(strict=True) so the fix cannot land silently.

THE FROZEN TEST SET HAS 20 DUPLICATE RECORDS. 385 records, 365 unique texts;
the 20 are byte-identical in text and spans, so effective n is 365 and the
bootstrap intervals are marginally optimistic. Measured impact is immaterial:
teacher F1 0.9482 -> 0.9482, student 0.5750 -> 0.5801, G1 ratio 0.6064 ->
0.6118, all far inside a +/-0.042 interval. Keeping the freeze, because
regenerating a test set after seeing results is precisely what the contract
forbids. Pinned at exactly 20 so it cannot drift.

The important negative: test has ZERO training leakage (0/385). Every
published number stands.

+++ forge/ci.py
@@ -57,7 +57,7 @@ class Interval:
         return (self.hi - self.lo) / 2
 
     def __str__(self) -> str:
-        pct = int(round((1 - self.alpha) * 100))
+        pct = round((1 - self.alpha) * 100)
         return f"{self.point:.4f} [{self.lo:.4f}, {self.hi:.4f}] {pct}% CI"
